feat: chat reaction notify, admin UM redesign, auth polish & progress fixes
- classroom chat: reaction notifications (bloc/page/body + backend service,
  Notification model, deep-link navigation)
- admin user management: redesigned details dialog, user card & action menu;
  bloc/repository/datasource updates
- auth: refreshed login/register/forgot/reset password & otp screens
- progress: report & stat-detail tweaks, examAttemptProgress + progressService
- speaking: expanded feedback page; writing task page fixes
- core: JWT interceptor refresh handling; app_en/vi l10n strings
- ci: leave auto-candidate changelog empty (app shows localized default)

+++ english_for_community/lib/core/entity/progress_summary_entity.dart
@@ -121,7 +121,7 @@ class CalloutEntity {
 class ProgressDetailEntity {
   final String id;
   final String title;
-  final int score; // Điểm số (ví dụ: 85% hoặc band score 6)
+  final double score; // Điểm số (85% hoặc band Viết 6.5 — giữ thập phân, format ở UI)
   final int duration; // Thời gian (phút)
   final String date; // ISO Date String
   final String type;
@@ -148,7 +148,7 @@ class ProgressDetailEntity {
       date: dateFromJson ?? 'N/A',
 
       // Sửa lỗi cho số (như đã thảo luận trước):
-      score: (json['score'] as num? ?? 0).toInt(),
+      score: (json['score'] as num? ?? 0).toDouble(),
       duration: (json['duration'] as num? ?? 0).toInt(),
 
       // Trường bổ sung
